Remove videoRecord class
The videoRecord class is no longer needed now that the functionality
has been moved into the cam-pipeline daemon. Some work is still needed
to re-implement raw recording modes.

+++ src/playbackwindow.cpp
@@ -328,7 +328,7 @@ void playbackWindow::checkForSaveDone()
 	}
 	else {
 		char tmp[64];
-		sprintf(tmp, "%.1ffps", (float)st.framerate);
+		sprintf(tmp, "%.1ffps", st.framerate);
 		ui->lblFrameRate->setText(tmp);
 	}
 }
